+++ R/misc.R
@@ -306,4 +306,3 @@ dummy.refs <- function(ref, item){
 }
 
 
-