+++ R/zzz.R
@@ -1,3 +1,3 @@
 .onLoad <- function(libname = find.package("childsds"), pkgname = "childsds"){
-    if(getRversion() >= "3.1.0") utils::globalVariables(c(".data", "group"))
+    if(getRversion() >= "3.1.0") utils::globalVariables(c(".data", "group","uchu.pal"))
 }